pyproject.toml
[tool.poetry]
name = "lachesis"
version = "0.1.0"
description = ""
authors = ["Pierre-Yves Gillier <github@pygillier.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
whitenoise = "^6.5.0"
django-redis = "^5.3.0"
django-environ = "^0.11.2"
django-model-utils = "^4.3.1"
crispy-bootstrap5 = "^0.7"
django-allauth = "^0.56.1"
django-bootstrap5 = "^23.3"
django-gravatar2 = "^1.4.4"
psycopg = { version = "^3.1.9", extras = ["binary"] }
django-unfold = "^0.10.0"

[tool.poetry.dev-dependencies]
mypy = "^1.4.1"
pytest = "==7.4.2"

flake8 = "==6.1.0"  # https://github.com/PyCQA/flake8
flake8-isort = "==6.0.0"  # https://github.com/gforcada/flake8-isort
coverage = "==7.3.1"  # https://github.com/nedbat/coveragepy
black = "==23.7.0"  # https://github.com/psf/black
djlint = "==1.32.1"  # https://github.com/Riverside-Healthcare/djLint
pylint-django = "==2.5.3"  # https://github.com/PyCQA/pylint-django
pre-commit = "==3.4.0"  # https://github.com/pre-commit/pre-commit

factory-boy= "==3.3.0"  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar = "==4.2.0"  # https://github.com/jazzband/django-debug-toolbar
django-extensions = "==3.2.3"  # https://github.com/django-extensions/django-extensions
django-coverage-plugin = "==3.1.0"  # https://github.com/nedbat/django_coverage_plugin
pytest-django = "==4.5.2"  # https://github.com/pytest-dev/pytest-django

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
